+++ modesolver/postprocess/__init__.py
@@ -1,3 +1,4 @@
 from .collocate import collocate
 from .poynting import poynting
 from .unfold import unfold
+from .scattering import gsm_step